Add slog helper
